Add API resources for backup and restore
BackupPolicy is a new CRD to define a backup policy.
This commit also enhance MySQLCluster to reference a BackupPolicy
and allows uesrs to specify how to restore data from a backup.

The status field of MySQLCluster has new fields for backup.

Read `docs/crd_backuppolicy.md` and `docs/crd_mysqlcluster.md`
for details.

This commit does not implement the controller logic necessary
for backup and restore.  They will be added with succeeding commits.